Fix activate.bat to properly detect VS2017
When both VS2017 Professional and the BuildTools are installed, `vswhere` returns two lines of output:

    >vswhere.exe -nologo -products * -version ^[15.0^,16.0^) -property installationPath
    C:\Program Files (x86)\Microsoft Visual Studio\2017\Professional
    C:\Program Files (x86)\Microsoft Visual Studio\2017\BuildTools

These are in random order, and only the last one is used to set `_VSINSTALLDIR`. If that happens to be the wrong one (Professional instead of BuildTools) then the SDK will not be detected, nothing will be added to the `PATH`, and compilations will fail with:

> 'cl.exe' is not recognized as an internal or external command, operable program or batch file.

This change ensures that we always find the BuildTools, even if Professional is also installed.